requirements.txt
fastapi
uvicorn
python-multipart
paddlepaddle
paddleocr
pdf2image
Pillow
numpy 